# By hucl
# Via hucl
* 'online_develop' of https://github.com/boxfish-project/boxfish-fishcard:
  将sku_id序列化到数据库中
  20120922上线
  分配教师报警代码提交

+++ boxfish-online-card/src/main/java/com/boxfishedu/workorder/web/view/course/ServiceWorkOrderCombination.java
@@ -88,7 +88,7 @@ public class ServiceWorkOrderCombination {
         for(WorkOrder workOrder : workOrderList) {
             RecommandCourseView recommandCourseView = recommendCourseViewMap.get(workOrder.getSeqNum());
             workOrder.initCourseInfo(recommandCourseView);
-            workOrder.setSkuId((long) CourseType2TeachingTypeService.courseType2TeachingType2(
+            workOrder.setSkuId( CourseType2TeachingTypeService.courseType2TeachingType2(
                     recommandCourseView.getCourseType(), TutorType.resolve(workOrder.getService().getTutorType())));
         }
         return this;
